+++ bootique-test/src/main/java/io/bootique/test/junit/BQModuleProviderChecker.java
@@ -1,12 +1,10 @@
 package io.bootique.test.junit;
 
-import com.google.inject.Module;
 import io.bootique.BQModuleProvider;
 import io.bootique.BQRuntime;
 import io.bootique.meta.module.ModuleMetadata;
 import io.bootique.meta.module.ModulesMetadata;
 
-import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.ServiceLoader;
@@ -15,9 +13,6 @@ import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
 import static java.util.stream.Collectors.counting;
-import static java.util.stream.Collectors.toList;
-import static org.hamcrest.core.IsCollectionContaining.hasItems;
-import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
@@ -48,31 +43,6 @@ public class BQModuleProviderChecker {
      */
     public static void testPresentInJar(Class<? extends BQModuleProvider> provider) {
         new BQModuleProviderChecker(provider).testPresentInJar();
-    }
-
-    /**
-     * Verifies that runtime contains expected modules.
-     *
-     * @param bqRuntime
-     * @param moduleList
-     * @since 0.25
-     */
-    public static void testModulesLoaded(BQRuntime bqRuntime, List<Class<? extends Module>> moduleList) {
-        final ModulesMetadata modulesMetadata = bqRuntime.getInstance(ModulesMetadata.class);
-
-        final List<String> actualModules = modulesMetadata
-                .getModules()
-                .stream()
-                .map(ModuleMetadata::getName)
-                .collect(toList());
-
-        final String[] expectedModules = moduleList
-                .stream()
-                .map(Class::getSimpleName)
-                .toArray(String[]::new);
-
-        // Using class names for checking module existing - weak.
-        assertThat(actualModules, hasItems(expectedModules));
     }
 
     /**